+++ lib/compass/Compass.h
@@ -19,12 +19,12 @@ class Compass {
 	virtual void setSmoothing(byte steps, bool adv)  = 0;
 	virtual void setCalibration(int x_min, int x_max, int y_min, int y_max, int z_min, int z_max)  = 0;
 
-	virtual int getX() = 0;
-	virtual int getY() = 0;
-	virtual int getZ() = 0;
-	virtual int getAzimuth() = 0;
-	virtual byte getBearing(int azimuth) = 0;
-	virtual void getDirection(char* myArray, int azimuth) = 0;
+	virtual float getX() = 0;
+	virtual float getY() = 0;
+	virtual float getZ() = 0;
+	virtual float getAzimuth() = 0;
+	virtual byte getBearing(float azimuth) = 0;
+	virtual void getDirection(char* myArray, float azimuth) = 0;
 protected:
 	const char _bearings[16][3] =  {
 		{' ', ' ', 'N'},
@@ -57,30 +57,30 @@ public:
 	virtual void setSmoothing(byte steps, bool adv) override;
 	virtual void setCalibration(int x_min, int x_max, int y_min, int y_max, int z_min, int z_max) override;
 
-	virtual int getX() override;
-	virtual int getY() override;
-	virtual int getZ() override;
-	virtual int getAzimuth() override;
-	virtual byte getBearing(int azimuth) override;
-	virtual void getDirection(char* myArray, int azimuth) override;
+	virtual float getX() override;
+	virtual float getY() override;
+	virtual float getZ() override;
+	virtual float getAzimuth() override;
+	virtual byte getBearing(float azimuth) override;
+	virtual void getDirection(char* myArray, float azimuth) override;
 
 
   protected:
     void _writeReg(byte reg,byte val);
-	int _get(int index);
+    float _get(int index);
 	bool _smoothUse = false;
 	byte _smoothSteps = 5;
 	bool _smoothAdvanced = false;
     byte _ADDR;
 	int _vRaw[3] = {0,0,0};
-	int _vHistory[10][3];
+	float _vHistory[10][3];
 	int _vScan = 0;
-	long _vTotals[3] = {0,0,0};
-	int _vSmooth[3] = {0,0,0};
+	float _vTotals[3] = {0,0,0};
+	float _vSmooth[3] = {0,0,0};
 	void _smoothing();
 	bool _calibrationUse = false;
 	int _vCalibration[3][2];
-	int _vCalibrated[3];
+    float _vCalibrated[3];
 	void _applyCalibration();
     
 };
@@ -100,12 +100,12 @@ public:
 	virtual void setSmoothing(byte steps, bool adv) { if(_compass!=nullptr) _compass->setSmoothing(steps, adv); };
 	virtual void setCalibration(int x_min, int x_max, int y_min, int y_max, int z_min, int z_max)  { if(_compass!=nullptr) _compass->setCalibration(x_min,x_max,y_min,y_max,z_min,z_max); };
 
-	virtual int getX() { if(_compass!=nullptr) return _compass->getX(); return 0; };
-	virtual int getY() { if(_compass!=nullptr) return _compass->getY(); return 0; };
-	virtual int getZ() { if(_compass!=nullptr) return _compass->getZ(); return 0; };
-	virtual int getAzimuth() { if(_compass!=nullptr) return _compass->getAzimuth(); return 0; };;
-	virtual byte getBearing(int azimuth) { if(_compass!=nullptr) return _compass->getBearing(azimuth); return 0; };
-	virtual void getDirection(char* myArray, int azimuth) { if(_compass!=nullptr) _compass->getDirection(myArray, azimuth); };
+	virtual float getX() { if(_compass!=nullptr) return _compass->getX(); return 0; };
+	virtual float getY() { if(_compass!=nullptr) return _compass->getY(); return 0; };
+	virtual float getZ() { if(_compass!=nullptr) return _compass->getZ(); return 0; };
+	virtual float getAzimuth() { if(_compass!=nullptr) return _compass->getAzimuth(); return 0; };;
+	virtual byte getBearing(float azimuth) { if(_compass!=nullptr) return _compass->getBearing(azimuth); return 0; };
+	virtual void getDirection(char* myArray, float azimuth) { if(_compass!=nullptr) _compass->getDirection(myArray, azimuth); };
 protected:
     Compass* _compass;
 };
